requirements.txt
scikit-learn
tensorflow_hub
tensorflow_datasets
matplotlib
Pillow
numpy
split-folders
